chore(xo-web/VM): move boot order setting from Disks to Advanced
See #1523
